requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# LangChain & AI
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.72
langchain-huggingface==0.3.0
langchain-groq==0.2.1

# Vector Database
chromadb==0.4.24

# ML & Embeddings
sentence-transformers==2.2.2
transformers==4.30.0
torch==2.0.1

# Groq API
groq==0.31.0

# Utilities
python-dotenv==1.0.0
pydantic==2.11.9
numpy==1.24.0